requirements.txt
requests
tabulate
python-editor
